Fix MSVC warning. DONT_BUILD.
git-svn-id: https://svn.eduke32.com/eduke32@5504 1a8010ca-5511-0410-912e-c29ae57300e0

+++ polymer/eduke32/source/actors.c
@@ -2757,7 +2757,7 @@ ACTOR_STATIC void Proj_MoveCustom(int32_t i)
             if (!m && (unsigned) s->owner < MAXSPRITES)
             {
                 cst = sprite[s->owner].cstat;
-                sprite[s->owner].cstat &= ~CSTAT_SPRITE_BLOCK;
+                sprite[s->owner].cstat &= (uint16_t)~CSTAT_SPRITE_BLOCK;
             }
 
             do
